Title

Title

Usage

grid_panels(
  panels_df,
  column_var,
  row_var,
  scales_x = c("free", "fixed_column", "fixed_row", "fixed_all", "fixed_combination"),
  scales_y = c("free", "fixed_column", "fixed_row", "fixed_all", "fixed_combination"),
  scales_y_coverage = c("free", "fixed_column", "fixed_row", "fixed_all",
    "fixed_combination"),
  panel_titles = FALSE,
  col_title_ratio = 30,
  row_title_ratio = 30,
  font_size = 18,
  new_y_label = NULL,
  new_x_label = NULL,
  title_size = 8
)

Arguments

panels_df

A dataframe from plot_seasonal_gam_panels.

column_var

1 dimensional character vector of variable defining the columns of the grid. Must be factor variable present in panels_df.

row_var

as column_var, but defines the rows of the grid

scales_x

Allows shared x limits for all panels within each column ("fixed_column"), or each row ("fixed_row"), each column x row combinations ("fixed_combination"; useful when more than one panel matches a row x column criterion). Defaults to allowing each panel to have its own x scale ("free").

scales_y

As scales_x, but for y limits. Is not applied to the optional coverage plots within each panel.

scales_y_coverage

As scales_y, but only applied to optional coverage plots.

panel_titles

Include all panel information in panel title even if redundant with rows/panels? Useful for debugging or screenshotting sections. Logical, defaults to false

col_title_ratio

How much bigger should content be than the column labels? Numeric, defaults to 30.

row_title_ratio

as col_title_ratio, but ratio of content to row labels.

font_size

Size of row and column labels. Numeric, defaults to 18.

new_y_label

Optional, replaces Y axis label of primary plots in panels (not data coverage histograms). character vector, defaults to NULL.

new_x_label

Optional, replaces X axis label of panels. Character vector, defaults to NULL.

title_size

size of panel title text. Numeric, defaults to 8.

Value

patchwork object